requirements.txt
wheel==0.37.1
apache-airflow==2.3.0
pandas==1.4.2
scikit-learn==1.0.2
mlflow==1.25.1
matplotlib==3.5.1